PESA-BACK/requirements.txt
amqp==1.4.9
anyjson==0.3.3
Babel==2.3.4
BeautifulSoup==3.2.1
beautifulsoup4==4.4.1
billiard==3.3.0.23
celery==3.1.23
Django==1.8
django-extensions==1.6.1
django-mailgun==0.8.0
django-nexmo==1.0.1
django-phonenumber-field==1.1.0
django-registration-redux==1.4
django-session-security==2.3.2
django-sslify==0.2.7
django-twilio==0.8.0
djangorestframework==3.3.3
html5lib==0.9999999
httplib2==0.9.2
jsonfield==1.0.3
kombu==3.0.35
M2Crypto==0.24.0
phonenumberslite==7.3.1
Pillow==3.2.0
psycopg2==2.6.1
pyasn1==0.1.9
pycrypto==2.6.1
python-keyczar==0.715
python-memcached==1.57
pytz==2016.4
PyYAML==3.11
requests==2.9.1
six==1.10.0
sorl-thumbnail==12.3
suds==0.4
twilio==5.4.0
xmltodict==0.10.1
